cache reservation

+++ BR/Services/ReservationService.cs
@@ -4,13 +4,16 @@ using BR.DTO.Reservations;
 using BR.EF;
 using BR.Models;
 using BR.Services.Interfaces;
+using BR.Utils;
 using BR.Utils.Notification;
 using Microsoft.Extensions.Caching.Distributed;
+using Microsoft.Extensions.Caching.Memory;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
 namespace BR.Services
@@ -19,15 +22,19 @@ namespace BR.Services
     {
         private readonly IAsyncRepository _repository;
         private readonly INotificationService _notificatinoService;
-        private readonly IDistributedCache _cache;
+        private readonly IDistributedCache _cacheDistributed;
+        private readonly IMemoryCache _cacheMemory;
+
 
         public ReservationService(IAsyncRepository repository,
             INotificationService notificationService,
-            IDistributedCache cache)
+            IDistributedCache cacheDistributed,
+            IMemoryCache cacheMemory)
         {
             _repository = repository;
             _notificatinoService = notificationService;
-            _cache = cache;
+            _cacheDistributed = cacheDistributed;
+            _cacheMemory = cacheMemory;
         }
 
         public async Task<ICollection<Reservation>> GetReservations(string identityUserId)
@@ -44,82 +51,118 @@ namespace BR.Services
         {
             return await _repository.GetReservation(id);
         }
-        public async Task<Reservation> AddNewReservation(NewReservationRequest newReservationRequest, string identityId)
+
+
+        public async Task<ServiceResponse<Reservation>> AddNewReservation(NewReservationRequest newReservationRequest, string identityId)
         {
             var user = await _repository.GetUser(identityId);
-            //var resState = await _repository.GetReservationState("idle");
-            var resDate = DateTime.ParseExact(newReservationRequest.StartDateTime, "dd/MM/yyyy HH:mm", CultureInfo.InvariantCulture);
-            var reservation = new Reservation()
-            {
-                UserId = user.Id,
-                ChildFree = newReservationRequest.IsChildFree,
-                GuestCount = newReservationRequest.GuestCount,
-                Comments = newReservationRequest.Comments,
-                ReservationDate = resDate,
-                ReservationStateId = null
-            };
-            reservation = await _repository.AddReservation(reservation);
-
-            foreach (var tableId in newReservationRequest.TableIds)
-            {
-                await _repository.AddTableReservation(reservation.Id, tableId);
-            }
             var client = await _repository.GetClientByTableId(newReservationRequest.TableIds.First());
-
-            // add data to redis
-            var interval = 15; 
-            ICollection<TableCurrentStateCacheData> tableStates = null;
-            var json = await _cache.GetStringAsync("tableStates");
-            if (json != null)
+            var resDate = DateTime.ParseExact(newReservationRequest.StartDateTime, "dd/MM/yyyy HH:mm", CultureInfo.InvariantCulture);
+            if (client.ReserveDurationAvg >= newReservationRequest.Duration)
             {
-                tableStates = JsonConvert.DeserializeObject<ICollection<TableCurrentStateCacheData>>(json);
+
+                var reservation = new Reservation()
+                {
+                    UserId = user.Id,
+                    ChildFree = newReservationRequest.IsChildFree,
+                    GuestCount = newReservationRequest.GuestCount,
+                    Comments = newReservationRequest.Comments,
+                    ReservationDate = resDate,
+                    Duration = newReservationRequest.Duration,
+                    ReservationStateId = null
+                };
+                reservation = await _repository.AddReservation(reservation);
+
+                foreach (var tableId in newReservationRequest.TableIds)
+                {
+                    await _repository.AddTableReservation(reservation.Id, tableId);
+                }
+
+
+                // add data to redis
+                await this.AddTableStateCacheData(resDate, newReservationRequest.Duration, newReservationRequest.TableIds);
+
+                if (client != null)
+                {
+                    var waiters = await _repository.GetWaitersByClientId(client.Id);
+                    if (waiters != null)
+                    {
+                        List<string> tags = new List<string>();
+                        foreach (var waiter in waiters)
+                        {
+                            var tokens = await _repository.GetTokens(waiter.IdentityId);
+                            foreach (var t in tokens)
+                            {
+                                tags.Add(t.NotificationTag);
+                            }
+                        }
+                        try
+                        {
+                            _notificatinoService.SendNotification("New reservation", MobilePlatform.gcm, "string", tags.ToArray());
+
+                        }
+                        catch
+                        {
+                            return new ServiceResponse<Reservation>(StatusCodeService.SendingNotificationError, reservation);
+                        }
+                    }
+                }
+                return new ServiceResponse<Reservation>(reservation);
             }
             else
             {
-                tableStates = new List<TableCurrentStateCacheData>();
-
+                return await SendReservationOnConfirmation(newReservationRequest, user.Id, client);
             }
-            foreach (var tableId in newReservationRequest.TableIds)
+        }
+
+        class ReservationConfirm
+        {
+            public int UserId { get; set; }
+            public string StartDaterTime { get; set; }
+            public int Duration { get; set; }
+            public ICollection<int> TableIds { get; set; }
+        }
+
+        public async Task<ServiceResponse<Reservation>> SendReservationOnConfirmation(NewReservationRequest newReservationRequest, int userId, Client client)
+        {
+            
+            // time for cache -> to appsettings
+
+
+            _cacheMemory.Set(userId, JsonConvert.SerializeObject(newReservationRequest), TimeSpan.FromMinutes(60));
+
+
+            var waiters = await _repository.GetWaitersByClientId(client.Id);
+            if (waiters != null)
             {
-
-                for (int i = 0; i < newReservationRequest.Duration; i += interval)
+                List<string> tags = new List<string>();
+                foreach (var waiter in waiters)
                 {
-                    tableStates.Add(new TableCurrentStateCacheData()
+                    var tokens = await _repository.GetTokens(waiter.IdentityId);
+                    foreach (var t in tokens)
                     {
-                        TableId = tableId,
-                        IsBusy = true,
-                        DateTime = resDate.AddMinutes(i)
-                    });
-                }
-            }
-
-            json = JsonConvert.SerializeObject(tableStates);
-
-            await _cache.SetStringAsync("tableStates", json);
-
-            //await _cache.SetStringAsync("tableStates", json, new DistributedCacheEntryOptions
-            //{
-            //    AbsoluteExpirationRelativeToNow = TimeSpan.FromSeconds(30)
-            //});
-
-            if (client != null)
-            {
-                var waiters = await _repository.GetWaitersByClientId(client.Id);
-                if (waiters != null)
-                {
-                    List<string> tags = new List<string>();
-                    foreach (var waiter in waiters)
-                    {
-                        var tokens = await _repository.GetTokens(waiter.IdentityId);
-                        foreach (var t in tokens)
-                        {
-                            tags.Add(t.NotificationTag);
-                        }
+                        tags.Add(t.NotificationTag);
                     }
-                    _notificatinoService.SendNotification("New reservation", MobilePlatform.gcm, "string", tags.ToArray());
+                }
+                try
+                {
+                    _notificatinoService.SendNotification("New reservation for confirmation", MobilePlatform.gcm, "string", tags.ToArray());
+                    return new ServiceResponse<Reservation>(StatusCodeService.SendOnConfirmation, null);
+                }
+                catch
+                {
+                    return new ServiceResponse<Reservation>(StatusCodeService.SendingNotificationError, null);
                 }
             }
-            return reservation;
+            else
+            {
+                return new ServiceResponse<Reservation>(StatusCodeService.Error, null);
+            }
+        }
+
+        public async Task<ServiceResponse<Reservation>> AddConfirmedReservation(ConfirmReservationRequest confirmRequest)
+        {
+
         }
 
         public async Task<Reservation> CancelReservation(int reservationId)
@@ -131,6 +174,14 @@ namespace BR.Services
                 return null;
             }
             reservation.ReservationStateId = resState.Id;
+            ICollection<int> tableIds = new List<int>();
+            foreach (var table in reservation.TableReservations)
+            {
+                tableIds.Add(table.TableId);
+            }
+
+            // remove redis data
+            await this.RemoveTableStateCacheData(reservation.ReservationDate, reservation.Duration, tableIds);
             return await _repository.UpdateReservation(reservation);
 
         }
@@ -145,6 +196,16 @@ namespace BR.Services
 
         public async Task ChangeTable(ChangeReservationTablesRequest changeRequest)
         {
+            var reservation = await _repository.GetReservation(changeRequest.ReservationId);
+            if (reservation is null)
+            {
+                return;
+            }
+            ICollection<int> tableIdsPrev = new List<int>();
+            foreach (var table in reservation.TableReservations)
+            {
+                tableIdsPrev.Add(table.TableId);
+            }
             await _repository.DeleteTableReservations(changeRequest.ReservationId);
 
             foreach (var tableId in changeRequest.TableIds)
@@ -154,6 +215,114 @@ namespace BR.Services
 
             // change redis data
 
+            await this.ChangeTableStateCacheData(reservation.ReservationDate, reservation.ReservationDate, reservation.Duration, reservation.Duration, tableIdsPrev, changeRequest.TableIds);
+
+
+        }
+
+
+
+        private async Task AddTableStateCacheData(DateTime timeStart, int duration, ICollection<int> tableIds)
+        {
+            var interval = 15;  // add to appsettings
+            ICollection<TableCurrentStateCacheData> tableStates = null;
+            var json = await _cacheDistributed.GetStringAsync("tableStates");
+            if (json != null)
+            {
+                tableStates = JsonConvert.DeserializeObject<ICollection<TableCurrentStateCacheData>>(json);
+            }
+            else
+            {
+                tableStates = new List<TableCurrentStateCacheData>();
+
+            }
+            foreach (var tableId in tableIds)
+            {
+
+                for (int i = 0; i < duration; i += interval)
+                {
+                    tableStates.Add(new TableCurrentStateCacheData()
+                    {
+                        TableId = tableId,
+                        DateTime = timeStart.AddMinutes(i)
+                    });
+                }
+            }
+
+            json = JsonConvert.SerializeObject(tableStates);
+
+            await _cacheDistributed.SetStringAsync("tableStates", json);
+        }
+
+        private async Task RemoveTableStateCacheData(DateTime timeStart, int duration, ICollection<int> tableIds)
+        {
+            var interval = 15;  // add to appsettings
+
+            var json = await _cacheDistributed.GetStringAsync("tableStates");
+
+            if (json is null)
+            {
+                return;
+            }
+            ICollection<TableCurrentStateCacheData> tableStates = JsonConvert.DeserializeObject<ICollection<TableCurrentStateCacheData>>(json);
+
+            foreach (var tableId in tableIds)
+            {
+                for (int i = 0; i < duration; i += interval)
+                {
+                    var tableState = tableStates.FirstOrDefault(t => t.TableId == tableId && t.DateTime == timeStart.AddMinutes(i));
+                    if (tableState != null)
+                    {
+                        tableStates.Remove(tableState);
+                    }
+                }
+            }
+
+            json = JsonConvert.SerializeObject(tableStates);
+
+            await _cacheDistributed.SetStringAsync("tableStates", json);
+        }
+
+        private async Task ChangeTableStateCacheData(DateTime timePrev, DateTime timeNew, int durationPrev, int durationNew, ICollection<int> tableIdsPrev, ICollection<int> tableIdsNew)
+        {
+            var interval = 15;  // add to appsettings
+            var json = await _cacheDistributed.GetStringAsync("tableStates");
+            if (json is null)
+            {
+                return;
+            }
+            ICollection<TableCurrentStateCacheData> tableStates = JsonConvert.DeserializeObject<ICollection<TableCurrentStateCacheData>>(json);
+
+            // delete previous states
+            foreach (var tableId in tableIdsPrev)
+            {
+                for (int i = 0; i < durationPrev; i += interval)
+                {
+                    var tableState = tableStates.FirstOrDefault(t => t.TableId == tableId && t.DateTime == timePrev.AddMinutes(i));
+                    if (tableState != null)
+                    {
+                        tableStates.Remove(tableState);
+                    }
+                }
+            }
+
+            // add new states
+            foreach (var tableId in tableIdsNew)
+            {
+
+                for (int i = 0; i < durationNew; i += interval)
+                {
+                    tableStates.Add(new TableCurrentStateCacheData()
+                    {
+                        TableId = tableId,
+                        DateTime = timeNew.AddMinutes(i)
+                    });
+                }
+            }
+
+            json = JsonConvert.SerializeObject(tableStates);
+
+            await _cacheDistributed.SetStringAsync("tableStates", json);
         }
     }
 }
